Delete image files when deleting or editing a product

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const { validationResult } = require("express-validator/check")
 
 const Product = require("../models/product")
+const fileHelper = require("../util/file")
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -114,7 +115,10 @@ exports.postEditProduct = (req, res, next) => {
         return res.redirect("/") //User doesn't have the right to edit product
       }
       product.title = title
-      if (image) product.imageUrl = image.path
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl)
+        product.imageUrl = image.path
+      }
       product.price = price
       product.description = description
       return product.save().then(response => {
@@ -148,7 +152,12 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.params.productId
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then(product => {
+      if (!product) return next(new Error("Product not found"))
+      fileHelper.deleteFile(product.imageUrl) // delete corresponding image file
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       res.redirect("/admin/products")
     })
